backend/backend/requirements.txt
# Core NIDS Dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
pydantic>=2.0.0

# Network capture and parsing
scapy>=2.5.0

# Windows notifications
winrt-Windows.UI.Notifications>=1.0.0; sys_platform == "win32"
win10toast>=0.9; sys_platform == "win32"

# Configuration and logging
pyyaml>=6.0
loguru>=0.7.0

# Optional performance optimization
numba>=0.58.0

# MATLAB file support
scipy>=1.10.0
h5py>=3.8.0

# API Server Dependencies
flask>=2.3.0
flask-cors>=4.0.0

# Development Dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Additional utilities
requests>=2.31.0
websockets>=11.0.0